Cập nhật chức năng lập lịch làm
fix schedule

+++ HRM_backend/src/main/java/com/project/hrm/Repositorys/ShiftRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Date;
 import java.util.List;
 
-public interface ShiftRepository extends JpaRepository<Shift, Integer> {
+public interface ShiftRepository extends JpaRepository<Shift, Long> {
 
     List<Shift> findAllByShiftType(ShiftType shiftType);
 
-    Shift findOneById(Integer id);
+    Shift findOneById(Long id);
 }
